[iter2] Secure Tour Routes ...

+++ controllers/tourControllers.js
@@ -3,8 +3,10 @@ const mongoose = require("mongoose");
 
 // GET /tours
 const getAllTours = async (req, res) => {
+  const user_id = req.user._id;
+
   try {
-    const tours = await Tour.find({}).sort({ createdAt: -1 });
+    const tours = await Tour.find({ user_id }).sort({ createdAt: -1 });
     res.status(200).json(tours);
   } catch (error) {
     res.status(500).json({ message: "Failed to retrieve tours" });
@@ -14,23 +16,26 @@ const getAllTours = async (req, res) => {
 // POST /tours
 const createTour = async (req, res) => {
   try {
-    const newTour = await Tour.create({ ...req.body });
+    const user_id = req.user._id;
+    const newTour = await Tour.create({ ...req.body, user_id });
     res.status(201).json(newTour);
   } catch (error) {
-    res.status(400).json({ message: "Failed to create tour", error: error.message });
+    res
+      .status(400)
+      .json({ message: "Failed to create tour", error: error.message });
   }
 };
 
 // GET /tours/:tourId
 const getTourById = async (req, res) => {
   const { tourId } = req.params;
+  const user_id = req.user._id;
 
   if (!mongoose.Types.ObjectId.isValid(tourId)) {
     return res.status(400).json({ message: "Invalid tour ID" });
   }
-
   try {
-    const tour = await Tour.findById(tourId);
+    const tour = await Tour.findById(tourId, { user_id });
     if (tour) {
       res.status(200).json(tour);
     } else {
@@ -44,6 +49,7 @@ const getTourById = async (req, res) => {
 // PUT /tours/:tourId
 const updateTour = async (req, res) => {
   const { tourId } = req.params;
+  const user_id = req.user._id;
 
   if (!mongoose.Types.ObjectId.isValid(tourId)) {
     return res.status(400).json({ message: "Invalid tour ID" });
@@ -51,9 +57,9 @@ const updateTour = async (req, res) => {
 
   try {
     const updatedTour = await Tour.findOneAndUpdate(
-      { _id: tourId },
+      { _id: tourId, user_id: user_id },
       { ...req.body },
-      { new: true }
+      { new: true },
     );
     if (updatedTour) {
       res.status(200).json(updatedTour);
@@ -68,13 +74,14 @@ const updateTour = async (req, res) => {
 // DELETE /tours/:tourId
 const deleteTour = async (req, res) => {
   const { tourId } = req.params;
+  const user_id = req.user._id;
 
   if (!mongoose.Types.ObjectId.isValid(tourId)) {
     return res.status(400).json({ message: "Invalid tour ID" });
   }
 
   try {
-    const deletedTour = await Tour.findOneAndDelete({ _id: tourId });
+    const deletedTour = await Tour.findOneAndDelete({ _id: tourId, user_id: user_id });
     if (deletedTour) {
       res.status(204).send(); // 204 No Content
     } else {
@@ -92,4 +99,3 @@ module.exports = {
   updateTour,
   deleteTour,
 };
-
